backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
psycopg2-binary==2.9.10
pyjwt==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
anthropic==0.79.0
requests==2.32.3
cryptography==44.0.0
gunicorn==22.0.0
gevent==24.11.1
Pillow==11.1.0
